Make extension for InteractiveUtils
