+++ Project.toml
@@ -11,7 +11,6 @@ DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 GraphsMatching = "c3af3a8c-b79e-4b01-bf44-c718d7e0e0d6"
 HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
-InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -31,12 +30,14 @@ SumTypes = "8e1ec7a9-0e02-4297-b0fe-6433085c89f2"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
 
 [weakdeps]
+InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 Tyler = "e170d443-b9d6-418a-9ee8-061e966341ef"
 
 [extensions]
 QuantumSavoryMakie = "Makie"
 QuantumSavoryTylerMakie = ["Tyler", "Makie"]
+QuantumSavoryInteractiveUtils = "InteractiveUtils"
 
 [compat]
 Combinatorics = "1"
