+++ src/QuantumSavory.jl
@@ -28,8 +28,6 @@ using SumTypes: @sum_type, isvariant, @cases
 import Combinatorics
 using Combinatorics: powerset
 
-import InteractiveUtils: subtypes
-
 import QuantumClifford
 import QuantumClifford: MixedDestabilizer
 
@@ -60,7 +58,7 @@ using QuantumSymbolics: I # to avoid ambiguity with LinearAlgebra.I
 
 export
     StateRef, RegRef, Register,
-    Qubit, Qumode, QuantumStateTrait,
+    Qubit, Qumode, QuantumStateTrait, available_slot_types,
     CliffordRepr, QuantumOpticsRepr, QuantumMCRepr,
     UseAsState, UseAsObservable, UseAsOperation,
     AbstractBackground,
@@ -82,7 +80,7 @@ export
     # quantumchannel.jl
     QuantumChannel,
     # backgrounds.jl
-    T1Decay, T2Dephasing, Depolarization, PauliNoise, AmplitudeDamping,
+    T1Decay, T2Dephasing, Depolarization, PauliNoise, AmplitudeDamping, available_background_types, constructor_metadata,
     # noninstant.jl
     AbstractNoninstantOperation, NonInstantGate, ConstantHamiltonianEvolution,
     # plots.jl
@@ -110,6 +108,12 @@ function __init__()
                 println(io, "\n`resourceplot_axis` requires the package `Makie`; please make sure `Makie` is installed and imported first.")
             elseif exc.f === generate_map
                 println(io, "\n`generate_map` requires the package `Tyler`; please make sure `Tyler` is installed and imported first.")
+            elseif exc.f === available_slot_types
+                println(io, "\n`available_slot_types` requires the package `InteractiveUtils`; please make sure `InteractiveUtils` is installed and imported first.")
+            elseif exc.f === available_background_types
+                println(io, "\n`available_background_types` requires the package `InteractiveUtils`; please make sure `InteractiveUtils` is installed and imported first.")
+            elseif exc.f === constructor_metadata
+                println(io, "\n`constructor_metadata` requires the package `InteractiveUtils`; please make sure `InteractiveUtils` is installed and imported first.")
             end
         end
     end
