+++ src/traits_and_defaults.jl
@@ -13,11 +13,7 @@ struct Qumode <: QuantumStateTrait end
 default_repr(::Qubit) = QuantumOpticsRepr()
 default_repr(::Qumode) = QuantumOpticsRepr()
 
+"""Display all available slot types in QuantumSavory along with their documentation.
 
-function available_slot_types() # TODO move this to an extension that loads when the InteractiveUtils is loaded
-    types = subtypes(QuantumStateTrait)
-
-    docs = [(type = T, doc = Base.Docs.doc(T)) for T in types] #TODO: edge case: no doc
-
-    return docs
-end
+The `InteractiveUtils` package must be installed and imported."""
+function available_slot_types end
